+++ packages/nostr_signaling/lib/nostr_signaling.dart
@@ -19,5 +19,4 @@ export 'src/implementations/event_callback.dart';
 export 'src/interfaces/i_event_callback.dart';
 export 'src/nostr_relay_list.dart';
 export 'src/nostr_signaling_factory.dart';
-export 'src/initial_point/initial_point_nostr_signaling.dart';
-export 'src/initial_point/initial_point_nostr_signaling_registry.dart';
+export 'src/initial_point/initial_point.dart';